refactor: login and signup moved to sequelize
