--- Backend/db/database_script.sql
@@ -1,9 +1,0 @@
--- CREATE DATABASE project_listings;
-
-CREATE TABLE users (
-    id SERIAL NOT NULL PRIMARY KEY,
-    firstName VARCHAR(50) NOT NULL,
-    lastName VARCHAR(50) NOT NULL,
-    email VARCHAR(150) NOT NULL,
-    password VARCHAR(150) NOT NULL
-);
