--- Backend/db/index.ts
@@ -1,10 +1,0 @@
-import { Pool } from "pg";
-
-const pool = new Pool({
-    user: process.env.DB_USER,
-    host: process.env.DB_HOST,
-    database: process.env.DB_NAME,
-    port: parseInt(process.env.DB_PORT || "5432"),
-});
-
-export default pool;
