+++ Backend/routes/auth/auth.ts
@@ -3,7 +3,7 @@ import passport from "passport";
 import signup from "../../controllers/auth/signup";
 import login from "../../controllers/auth/login";
 
-export type User = {
+export type UserType = {
     id?: number;
     firstName: string;
     lastName: string;
